pyproject.toml
[tool.poetry]
name = "parking-API"
version = "0.1.0"
description = "\"API for booking parking\""
authors = ["Michal99Zajac <michal.99.zajac@gmail.com>"]
readme = "README.md"
packages = [{include = "manage.py"}, {include = "src"}, {include = "migrations"}]
exclude = [
  { path = "src/db/alembic" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.88.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
pydantic = {extras = ["email"], version = "^1.10.4"}
alembic = "^1.9.0"
typer = {extras = ["all"], version = "^0.7.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
sqlalchemy = "^2.0.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
mypy = "^0.991"
flake8 = "^6.0.0"
autoflake = "^2.0.1"

[tool.isort]
line_length = 100

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core"]
